Add backed support for a 'special' swatch
A new swatch is to be supplied, called 'special' this will be used on the optional fronts that straddle "Top Stories". These fronts will be used for special reports, huge news, elections - that sort of thing. The fronts tool will be able to supply once: https://github.com/guardian/facia-tool/pull/989 is merged.

I don't yet have a definition of the swatch values from @blongden73