Re-enable landscape orientation.
From the external beta we have had lots of feedback from existing users that landscape orientation is very important to them.
For this reason we have decided to add it back (acknowledging it's currently not perfect). This will help us raise defects against it during the beta,
and understand just how much work we have here.