Interviews now pull the article template
Interviews will have their own template in the future, but they don't right now. 
They've been using the immersive template, but that creates image crop issues because these are not checked by subs on composer. 
This fix makes interviews pull the article template, but their placing in the string has been kept where it was before.